+++ packages/reactabular-tree/src/pack.js
@@ -1,32 +1,59 @@
-const packTree = ({
+import cloneDeep from 'lodash/cloneDeep';
+import find from 'lodash/find';
+import omit from 'lodash/omit';
+
+const pack = ({
+  idField = 'id',
   parentField = 'parent',
   childrenField = 'children'
 } = {}) => (rows) => {
-  const ret = [];
-  let pack = [];
-  let previousParent;
+  if (!Array.isArray(rows)) {
+    return [];
+  }
 
-  rows.forEach((row) => {
-    if (typeof row[parentField] !== 'undefined') {
-      pack.push(row);
+  if (!rows.length) {
+    return rows;
+  }
+
+  const ret = [];
+  let previousParents;
+
+  // Clone deep to guarantee immutability (implementation relies on
+  // references)
+  cloneDeep(rows).forEach((row) => {
+    const rowParent = row[parentField];
+
+    if (typeof rowParent !== 'undefined') {
+      // Check if the parent field matches the accumulated potential parents
+      const foundParent = find(previousParents, { [idField]: rowParent });
+
+      if (foundParent) {
+        // Found the matching parent, set up children + push
+        if (!foundParent[childrenField]) {
+          foundParent[childrenField] = [];
+        }
+
+        foundParent[childrenField].push(omit(row, parentField));
+
+        previousParents.push(
+          last(foundParent[childrenField])
+        );
+      } else {
+        console.warn('Invalid parent id', row, previousParents); // eslint-disable-line no-console
+      }
     } else {
       ret.push(row);
 
-      if (previousParent && pack && pack.length) {
-        previousParent[childrenField] = pack;
-
-        pack = [];
-      }
-
-      previousParent = row;
+      // No parents, flush lookup
+      previousParents = [row];
     }
   });
-
-  if (pack.length && previousParent) {
-    previousParent[childrenField] = pack;
-  }
 
   return ret;
 };
 
-export default packTree;
+function last(arr) {
+  return arr[arr.length - 1];
+}
+
+export default pack;